+++ example/android/app/src/main/kotlin/com/example/ohlcv_chart_example/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.example.ohlcv_chart_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()